fix for 64 bit hosts
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@24530 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/inadyn/src/os_windows.c
@@ -215,7 +215,7 @@ RC_TYPE os_shell_execute(char * p_cmd)
 	return rc; 
 }
 
-static unsigned long os_get_inet_addr(char* addr)
+static uint32_t os_get_inet_addr(char* addr)
 {
     unsigned long b3;
     unsigned long b2;
